Separated game setup into its own class for more extensibility.
